+++ docs/src/components.js
@@ -1,0 +1,7 @@
+import { GoMarkGithub } from 'react-icons/go'
+import { FaTwitter } from 'react-icons/fa'
+
+export default {
+	GithubIcon: GoMarkGithub,
+	TwitterIcon: FaTwitter,
+}