+++ example/src/Examples/NestedStates/components/Graph.tsx
@@ -5,11 +5,11 @@ import { useStateDesigner } from "state-designer"
 type Props = {}
 
 const Graph: React.FC<Props> = ({ children }) => {
-  const { getGraph } = useStateDesigner(state)
+  const { graph } = useStateDesigner(state)
 
   return (
     <pre>
-      <code>{JSON.stringify(getGraph(), null, 2)}</code>
+      <code>{JSON.stringify(graph, null, 2)}</code>
     </pre>
   )
 }