+++ example/src/Examples/Switch/index.tsx
@@ -7,8 +7,6 @@ export interface Props {}
 const Switch: React.FC<Props> = ({ children }) => {
   const { data, send, isIn } = useStateDesigner(state)
 
-  // React.useEffect(() => console.log("changed"), [data])
-
   return (
     <div className="example">
       <h2>Switch</h2>